SAK-11245 http://jira.sakaiproject.org/jira/browse/SAK-11245
Integration test working, but not optimizing the central index yet.
Neesds some more work. almost there.



git-svn-id: https://source.sakaiproject.org/svn/search/trunk@35644 66ffb92e-73f9-0310-93c1-f5514f145a0a

+++ search/search-impl/impl/src/java/org/sakaiproject/search/journal/impl/SharedFilesystemJournalStorage.java
@@ -120,15 +120,18 @@ public class SharedFilesystemJournalStorage implements JournalStorage
 	 */
 	public void rollbackSave(JournalStorageState jss)
 	{
-		File journalZip = ((JournalStorageStateImpl) jss).journalZip;
-		File tmpZip = ((JournalStorageStateImpl) jss).tmpZip;
-		if (tmpZip != null && tmpZip.exists())
+		if (jss != null)
 		{
-			tmpZip.delete();
-		}
-		if (journalZip != null && journalZip.exists())
-		{
-			journalZip.delete();
+			File journalZip = ((JournalStorageStateImpl) jss).journalZip;
+			File tmpZip = ((JournalStorageStateImpl) jss).tmpZip;
+			if (tmpZip != null && tmpZip.exists())
+			{
+				tmpZip.delete();
+			}
+			if (journalZip != null && journalZip.exists())
+			{
+				journalZip.delete();
+			}
 		}
 	}
